requirements-m1.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
llama-index==0.9.39
lxml==5.1.0
marshmallow==3.20.2
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
packaging==23.2
pandas==2.2.0
peewee==3.17.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.36
